refactor: move all code to server and separate to 3 small parts
