+++ Content.Server/_Scp/LightFlicking/Components/LightFlickingComponent.cs
@@ -1,12 +1,9 @@
-﻿using Robust.Shared.GameStates;
+﻿namespace Content.Server._Scp.LightFlicking.Components;
 
-namespace Content.Shared._Scp.LightFlicking;
-
-[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
+[RegisterComponent]
 public sealed partial class LightFlickingComponent : Component
 {
-    [AutoNetworkedField, ViewVariables]
-    public bool Enabled;
+    [ViewVariables] public bool Enabled;
 
     [ViewVariables] public TimeSpan? NextFlickStartChanceTime = null;
     [ViewVariables] public TimeSpan NextFlickTime;
