+++ Content.Server/_Scp/LightFlicking/Components/MalfunctionLightComponent.cs
@@ -1,0 +1,7 @@
+﻿namespace Content.Server._Scp.LightFlicking.Components;
+
+[RegisterComponent]
+public sealed partial class MalfunctionLightComponent : Component
+{
+
+}
