--- Content.Shared/_Scp/LightFlicking/MalfunctionLightComponent.cs
@@ -1,9 +1,0 @@
-﻿using Robust.Shared.GameStates;
-
-namespace Content.Shared._Scp.LightFlicking;
-
-[RegisterComponent, NetworkedComponent]
-public sealed partial class MalfunctionLightComponent : Component
-{
-
-}
